requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0
scapy==2.5.0
netifaces==0.11.0
dpkt==1.9.8
google-generativeai==0.3.2
huggingface_hub
requests 